requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dotenv==0.21.1